+++ ididi/_type_resolve.py
@@ -7,9 +7,22 @@ import contextlib
 import inspect
 import sys
 import types
-import typing as ty
+from typing import (
+    Any,
+    AsyncContextManager,
+    Awaitable,
+    Callable,
+    ContextManager,
+    ForwardRef,
+    Protocol,
+    TypeVar,
+    Union,
+    cast,
+    get_args,
+    get_origin,
+)
 
-import typing_extensions as tyex
+from typing_extensions import ParamSpec, TypeGuard
 
 from ._itypes import AsyncClosable, Closable
 from .errors import (
@@ -19,18 +32,18 @@ from .errors import (
 )
 from .utils.typing_utils import T, eval_type, get_full_typed_signature, is_builtin_type
 
-SyncResource = ty.Union[ty.ContextManager[ty.Any], Closable]
-AsyncResource = ty.Union[ty.AsyncContextManager[ty.Any], AsyncClosable]
+SyncResource = Union[ContextManager[Any], Closable]
+AsyncResource = Union[AsyncContextManager[Any], AsyncClosable]
 
 
-P = tyex.ParamSpec("P")
+P = ParamSpec("P")
 
 
-class EmptyInitProtocol(ty.Protocol): ...
+class EmptyInitProtocol(Protocol): ...
 
 
 def get_typed_signature(
-    call: ty.Callable[..., T],
+    call: Callable[..., T],
     check_return: bool = False,
 ) -> inspect.Signature:
     """
@@ -47,16 +60,16 @@ def resolve_factory_return(sig_return: type[T]) -> type[T]:
     """
     Get dependent type from a factory return
     """
-    if ty.get_origin(sig_return) in (
+    if get_origin(sig_return) in (
         collections.abc.AsyncGenerator,
         collections.abc.Generator,
     ):
-        dependent, *_ = ty.get_args(sig_return)
+        dependent, *_ = get_args(sig_return)
         return dependent
     return sig_return
 
 
-def is_unresolved_type(t: ty.Any) -> bool:
+def is_unresolved_type(t: Any) -> bool:
     """
     Types that are not resolved at type resolving.
     Builtin types are unresolvable, but we might have other cases.
@@ -70,33 +83,31 @@ def is_unresolved_type(t: ty.Any) -> bool:
     return is_builtin_type(t)
 
 
-def is_context_manager(t: T) -> tyex.TypeGuard[ty.ContextManager[T]]:
+def is_context_manager(t: T) -> TypeGuard[ContextManager[T]]:
     return isinstance(t, contextlib.AbstractContextManager)
 
 
-def is_async_context_manager(t: T) -> tyex.TypeGuard[ty.AsyncContextManager[T]]:
+def is_async_context_manager(t: T) -> TypeGuard[AsyncContextManager[T]]:
     return isinstance(t, contextlib.AbstractAsyncContextManager)
 
 
-def is_class_or_method(obj: ty.Any) -> bool:
+def is_class_or_method(obj: Any) -> bool:
     return isinstance(obj, (type, types.MethodType, classmethod))
 
 
 def is_class(
-    obj: ty.Union[type[T], ty.Callable[..., ty.Union[T, ty.Awaitable[T]]]]
-) -> tyex.TypeGuard[type[T]]:
+    obj: Union[type[T], Callable[..., Union[T, Awaitable[T]]]]
+) -> TypeGuard[type[T]]:
     """
     check if obj is a class, since inspect only checks if obj is a class type.
     """
-    origin = ty.get_origin(obj) or obj
+    origin = get_origin(obj) or obj
     is_type = isinstance(origin, type)
     is_generic_alias = isinstance(obj, types.GenericAlias)
     return is_type or is_generic_alias
 
 
-def is_function(
-    obj: ty.Union[type[T], ty.Callable[P, T]]
-) -> tyex.TypeGuard[ty.Callable[P, T]]:
+def is_function(obj: Union[type[T], Callable[P, T]]) -> TypeGuard[Callable[P, T]]:
     return isinstance(obj, types.FunctionType)
 
 
@@ -114,8 +125,8 @@ def is_class_with_empty_init(cls: type) -> bool:
 
 
 def resolve_forwardref(
-    dependent: ty.Union[type, ty.Callable[..., ty.Any]], ref: ty.ForwardRef
-) -> ty.Any:
+    dependent: Union[type, Callable[..., Any]], ref: ForwardRef
+) -> Any:
     if is_function(dependent):
         globalvs = dependent.__globals__
     else:
@@ -127,30 +138,30 @@ def resolve_forwardref(
         raise ForwardReferenceNotFoundError(ref) from e
 
 
-def resolve_annotation(annotation: ty.Any) -> type:
-    origin = ty.get_origin(annotation) or annotation
+def resolve_annotation(annotation: Any) -> type:
+    origin = get_origin(annotation) or annotation
 
-    if isinstance(annotation, ty.TypeVar):
+    if isinstance(annotation, TypeVar):
         raise GenericDependencyNotSupportedError(annotation)
 
     # === Solvable dependency, NOTE: order matters!===
 
     if sys.version_info >= (3, 10):
-        union_meta = (types.UnionType, ty.Union)
+        union_meta = (types.UnionType, Union)
     else:
-        union_meta = (ty.Union,)
+        union_meta = (Union,)
 
     if origin in union_meta:
-        union_types = ty.get_args(annotation)
+        union_types = get_args(annotation)
         # we don't care which type is correct, it would be provided by the factory
         return union_types[0]
     return origin
 
 
 def get_bases(dependent: type) -> tuple[type, ...]:
-    if issubclass(dependent, ty.Protocol):
-        # -3 excludes ty.Protocol, ty.Gener, object
-        bases = ty.cast(type, dependent).__mro__[1:-3]
+    if issubclass(dependent, Protocol):
+        # -3 excludes Protocol, Gener, object
+        bases = cast(type, dependent).__mro__[1:-3]
     else:
         bases = dependent.__mro__[1:-1]
     return bases