+++ ididi/api.py
@@ -1,24 +1,22 @@
-import typing as ty
+from typing import Any, Awaitable, Callable, Union, overload
 
 from .graph import DependencyGraph as DependencyGraph
 from .utils.typing_utils import P, T
 
 
-def entry(
-    func: ty.Callable[P, T]
-) -> ty.Union[ty.Callable[..., T], ty.Callable[..., ty.Awaitable[T]]]:
+def entry(func: Callable[P, T]) -> Union[Callable[..., T], Callable[..., Awaitable[T]]]:
     dg = DependencyGraph()
     return dg.entry(func)
 
 
-@ty.overload
-def resolve(dep: ty.Callable[P, T], /) -> T: ...
+@overload
+def resolve(dep: Callable[P, T], /) -> T: ...
 
 
-@ty.overload
-def resolve(dep: ty.Callable[P, T], /, *args: P.args, **overrides: P.kwargs) -> T: ...
+@overload
+def resolve(dep: Callable[P, T], /, *args: P.args, **overrides: P.kwargs) -> T: ...
 
 
-def resolve(dep: ty.Callable[P, T], /, **overrides: ty.Any) -> T:
+def resolve(dep: Callable[P, T], /, **overrides: Any) -> T:
     dg = DependencyGraph()
     return dg.resolve(dep, **overrides)