+++ ididi/utils/typing_utils.py
@@ -1,37 +1,36 @@
 import inspect
-import typing as ty
+from typing import Any, Callable, ForwardRef, Mapping, TypeVar, Union
 from typing import _eval_type as ty_eval_type  # type: ignore
+from typing import cast
 
-import typing_extensions as tyex
+from typing_extensions import ParamSpec, TypeGuard
 
-T = ty.TypeVar("T")
-R = ty.TypeVar("R")
-P = tyex.ParamSpec("P")
+T = TypeVar("T")
+R = TypeVar("R")
+P = ParamSpec("P")
 
-PrimitiveBuiltins = type[ty.Union[int, float, complex, str, bool, bytes, bytearray]]
+PrimitiveBuiltins = type[Union[int, float, complex, str, bool, bytes, bytearray]]
 ContainerBuiltins = type[
-    ty.Union[list[T], tuple[T, ...], dict[ty.Any, T], set[T], frozenset[T]]
+    Union[list[T], tuple[T, ...], dict[Any, T], set[T], frozenset[T]]
 ]
 BuiltinSingleton = type[None]
 
 
-def is_builtin_primitive(t: ty.Any) -> tyex.TypeGuard[PrimitiveBuiltins]:
+def is_builtin_primitive(t: Any) -> TypeGuard[PrimitiveBuiltins]:
     return t in {int, float, complex, str, bool, bytes, bytearray, type}
 
 
-def is_builtin_container(t: ty.Any) -> tyex.TypeGuard[ContainerBuiltins[ty.Any]]:
+def is_builtin_container(t: Any) -> TypeGuard[ContainerBuiltins[Any]]:
     return t in {list, tuple, dict, set, frozenset}
 
 
-def is_builtin_singleton(t: ty.Any) -> tyex.TypeGuard[BuiltinSingleton]:
+def is_builtin_singleton(t: Any) -> TypeGuard[BuiltinSingleton]:
     return t is None
 
 
 def is_builtin_type(
-    t: ty.Any,
-) -> tyex.TypeGuard[
-    ty.Union[PrimitiveBuiltins, ContainerBuiltins[ty.Any], BuiltinSingleton]
-]:
+    t: Any,
+) -> TypeGuard[Union[PrimitiveBuiltins, ContainerBuiltins[Any], BuiltinSingleton]]:
     """
     Builtin types are ignored at type resolving.
     It must be provided by default value, or use a factory to override it.
@@ -42,16 +41,16 @@ def is_builtin_type(
     is_primitive = is_builtin_primitive(t)
     is_container = is_builtin_container(t)
     is_singleton = is_builtin_singleton(t)
-    return is_primitive or is_container or is_singleton or (t is ty.Any)
+    return is_primitive or is_container or is_singleton or (t is Any)
 
 
 def eval_type(
-    value: ty.ForwardRef,
-    globalns: ty.Union[dict[str, ty.Any], None] = None,
-    localns: ty.Union[ty.Mapping[str, ty.Any], None] = None,
+    value: ForwardRef,
+    globalns: Union[dict[str, Any], None] = None,
+    localns: Union[Mapping[str, Any], None] = None,
     *,
     lenient: bool = False,
-) -> ty.Any:
+) -> Any:
     """
     # NOTE: copy from pydantic, credit to them.
     Evaluate the annotation using the provided namespaces.
@@ -65,21 +64,21 @@ def eval_type(
     """
 
     try:
-        return ty.cast(type[ty.Any], ty_eval_type(value, globalns, localns))
+        return cast(type[Any], ty_eval_type(value, globalns, localns))
     except NameError:
         if not lenient:
             raise
         return value
 
 
-def get_typed_annotation(annotation: ty.Any, globalns: dict[str, ty.Any]) -> ty.Any:
+def get_typed_annotation(annotation: Any, globalns: dict[str, Any]) -> Any:
     if isinstance(annotation, str):
-        annotation = ty.ForwardRef(annotation, is_argument=False)
+        annotation = ForwardRef(annotation, is_argument=False)
         annotation = eval_type(annotation, globalns, globalns, lenient=True)
     return annotation
 
 
-def get_typed_params(call: ty.Callable[..., T]) -> list[inspect.Parameter]:
+def get_typed_params(call: Callable[..., T]) -> list[inspect.Parameter]:
     signature = inspect.signature(call)
     globalns = getattr(call, "__globals__", {})
     globalns.pop("copyright", None)  #
@@ -95,12 +94,12 @@ def get_typed_params(call: ty.Callable[..., T]) -> list[inspect.Parameter]:
     return typed_params
 
 
-def get_full_typed_signature(call: ty.Callable[..., T]) -> inspect.Signature:
+def get_full_typed_signature(call: Callable[..., T]) -> inspect.Signature:
     """
     Get a full typed signature from a callable.
     check_return: bool
 
-    if check_return is True, raise MissingReturnTypeError if the return type is inspect.Signature.empty.
+    if check_return is True, raise MissingReturnTypeError if the return type is inspect.Signature.emp
     """
     signature = inspect.signature(call)
     globalns = getattr(call, "__globals__", {})